+++ app/services/PostService.scala
@@ -6,8 +6,7 @@ import utils.EStatus
 import utils.EStatus.EStatus
 
 import javax.inject.Inject
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Future}
 
 class PostService @Inject()(
   dao: PostDao,
@@ -32,26 +31,26 @@ class PostService @Inject()(
     dao.all()
   }
 
-  def getByFriendIds(id: Long): Future[Seq[PostDisplayDto]] = {
-    // TODO: write a concurrent version of this method
-    val friends = Await.result(befriendsService.getFriendsByUserId(id), Duration.Inf);
-    val friendIds = friends.map(_.id);
-    val posts = friendIds.flatMap(id => Await.result(dao.getByOwnerId(id.get), Duration.Inf));
 
-    val res = posts.map {
-      post => {
-        val numLikes = Await.result(postReactionDao.getLikesByPostId(post.id.get), Duration.Inf).size
-        PostDisplayDto(
-          post.id,
-          post.title,
-          post.content,
-          post.dateTime.toLocalDateTime,
-          friends.find(_.id.get == post.ownerId).get,
-          numLikes)
+  def getByFriendIds(id: Long): Future[Seq[PostDisplayDto]] = {
+    for {
+      friends <- befriendsService.getFriendsByUserId(id)
+      posts <- Future.sequence(friends.map(_.id).map(id => dao.getByOwnerId(id.get))).map(_.flatten)
+      likes <- postReactionDao.allLikes()
+    } yield {
+      posts.map { post =>
+        {
+          PostDisplayDto(
+            post.id,
+            post.title,
+            post.content,
+            post.dateTime.toLocalDateTime,
+            friends.find(_.id.get == post.ownerId).get,
+            likes.filter(_.postId == post.id.get).size
+          )
+        }
       }
     }
-
-    Future { res }
   }
 
 
@@ -61,18 +60,24 @@ class PostService @Inject()(
 
 
   def getByOwnerId(id: Long): Future[Seq[PostDisplayDto]] = {
-    // TODO: write a concurrent version of this method
-    val owner = Await.result(userService.get(id), Duration.Inf);
-    val posts = Await.result(dao.getByOwnerId(id), Duration.Inf);
-
-    val res = posts.map {
-      post => {
-        val numLikes = Await.result(postReactionDao.getLikesByPostId(post.id.get), Duration.Inf).size
-        PostDisplayDto(post.id, post.title, post.content, post.dateTime.toLocalDateTime, owner.get, numLikes)
+    for {
+      owner <- userService.get(id)
+      posts <- dao.getByOwnerId(id)
+      likes <- postReactionDao.allLikes()
+    } yield {
+      posts.map { post =>
+        {
+          PostDisplayDto(
+            post.id,
+            post.title,
+            post.content,
+            post.dateTime.toLocalDateTime,
+            owner.get,
+            likes.filter(_.postId == post.id.get).size
+          )
+        }
       }
     }
-
-    Future { res }
   }
 
 
@@ -89,40 +94,35 @@ class PostService @Inject()(
   /********
    * POST *
    ********/
-  def create(postDto: PostInsertDto): Future[EStatus] = {
-    userService.existsId(postDto.ownerId).map {
-      case true =>
-        dao.insert(postDto)
-        EStatus.Success
-      case false =>
-        EStatus.Failure
-    }
+  def create(postInsertDto: PostInsertDto): Future[Option[Post]] = {
+    for {
+      _ <- dao.insert(postInsertDto)
+      s <- dao.num()
+      posts <- getAll()
+    } yield posts.sortBy(_.id).drop(s - 1).find(p => true)
   }
 
 
-  def update(post: Post): Future[EStatus] = {
-    dao.existsAndOwnedBy(post.id.get, post.ownerId).map {
-      case true =>
-        dao.update(post)
-        EStatus.Success
-      case false =>
-        EStatus.Failure
-    }
+  def update(post: Post): Future[Option[Post]] = {
+    for {
+      _ <- dao.update(post)
+      post <- get(post.id.get)
+    } yield post
   }
 
 
   def delete(id: Long): Future[EStatus] = {
-
-    dao.exists(id).map {
-      case true =>
-        // TODO: write a concurrent version of this method
-        Await.result(commentDao.getByPostId(id), Duration.Inf).map(c => commentDao.delete(c.id.get))
-        Await.result(postReactionDao.getByPostId(id), Duration.Inf).map(p => postReactionDao.delete(p.userId, p.postId))
+    for {
+      b <- dao.exists(id)
+      _ <- commentDao.deleteByPostId(id)
+      _ <- postReactionDao.deleteByPostId(id)
+    } yield {
+      if (b) {
         dao.delete(id)
-
         EStatus.Success
-      case false =>
+      } else {
         EStatus.Failure
+      }
     }
   }
